backend/pyproject.toml
[project]
name = "yakkum"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "PyMySQL>=1.1.0",
    "python-dotenv>=1.1.0",
    "whitenoise>=6.6.0",
]
